Add --index-only support with read_tree_trivial merges, kind of
This almost works.  It creates the correct merge commit, updates the
branch, but then if the repo is bare it leaves the index as it was before
the merge (and if the repo is non-bare it updates the index).  I'm
totally lost as to why the testcase behaves differently in the bare and
non-bare cases.