diffcore-rename: No point trying to find a match better than exact
diffcore_rename() had a line of code to avoid re-checking for renames of
paths who already had an exact match.  Sources, however, were re-checked
because we wanted to allow the possibility of detecting copies.  But if
copy detection isn't turned on, then this merely amounts to attempting
to find a better-than-exact match, which naturally ends up being an
expensive no-op.

In a large repository (~50k files, about 60% of which was java) that had
a number of nearly toplevel directories involved in renames, this cut
the time necessary for a cherry-pick down by about 50% (from around 9
minutes to 4.5 minutes).

Signed-off-by: Elijah Newren <newren@palantir.com>