Sweet idea:
If a file is unmodified on one side of history (no content changes and no
name change) and is renamed on the other side, then the correct merge
result is to take both the file name and the file contents of the renamed
file.

Note that if no rename detection is done, then this will appear to the
merge machinery as two files: one that was unmodified on one side of
history and deleted on the other (thus the merge should delete it), and
one which was newly added on one side of history (thus the merge should
include it).

In other words, rename detection is a waste of time for files that were
unmodified on the other side of history.

get_unmerged() would include all these paths anyway; see
Documentation/technical/trival-merge.txt -- unmodified + deleted is a
no merge candidate.  So we need to post-process that list by removing
all entries that are unmodified on both sides.  In fact, we can also
remove entries where the path exists on all three sides even if it's
different in all three locations (though the rename machine would have
ignored such paths anyway).

The only problem with this idea is directory renames, because we do need
to detect those in order to get files into the right location.  But what
if we only based those on 100% file renames?   ....   :-)

Secondary idea:

We could still have lots of paths with changes that we need to do rename
detection on due to directory renames.  But we could add a codepath in
diffcore_rename(), between the simple finding of exact matches and the
exhaustive search to find best possible matches, which (a) checks for
directory renames on our side of history (based on exact renames and
absence of any other paths inside the given directory), (b) for each path:
(b1) checks if directory un-renaming is applicable to this path, (b2)
checks if directory-unrenamed location exists in src, (b3) if so checks if
dst and src are similar enough to meet minimum threshold score.  If so,
record the rename pair (like the exact rename detection does) and avoid
using the source or destination in the exhaustive search.