merge-ort: basic new merge strategy
Copy functions from merge-recursive to implement a basic do-nothing
merge strategy.  The eventual goal is to be a drop-in replacement for
merge-recursive.c, but currently there is no actual merge logic beyond
calling unpack_trees(), hoping it handles everything, and if it does,
writing the merge out.

Note that even the functions which follow the "End copy-paste static
functions from merge-recursive.c" are still almost exact copy-paste,
in order to allow me to rebase more easily.  The only differences are:
  * The logic in merge_trees_internal() for handling the unmerged
    entries has been gutted; this will be re-added later.
  * merge_trees() and merge_recursive() were renamed to avoid
    conflicting symbols so that the build can succeed.

Signed-off-by: Elijah Newren <newren@gmail.com>