t6022, t6046: test expected behavior instead of testing a proxy for it
In t6022, we were testing for file being overwritten or not based on an
output message instead of checking for the file being overwritten, which
we can check via mtime updates.  Check for mtime updates instead.

In t6046, we were largely checking for both the expected behavior and a
proxy for it, which is unnecessary.  The calls to test-tool also were a
bit cryptic.  Make them a little clearer.

Signed-off-by: Elijah Newren <newren@gmail.com>