t6043: swap ruling for testcase 8e
I could mimic the same functionality if I modified
handle_directory_level_conflicts() to instead:

  * Remove rename/rename(1to1) cases from both side[12]_dir_renames
  * Call remove_invalid_dir_renames()
  * Remove rename/rename(1to2) cases from both side[12]_dir_renames

Because separating the 1to1 and the 1to2 cases with putting the call
to remove_invalid_dir_renames() inbetween makes cases 6b and 8e work
as they did in merge-recursive.  But I kind of like the idea of having
both side rename from the same location imply no directory rename
detection, especially since I later want to optimize and say that there
are no directory renames for a source directory unless the dirmask is
either 3 or 5 (in particular, dirmask of 1 which correspond to both of
these cases then gets thrown out).