diffcore-rename: initial stab at improved basename guessing
Still has two FIXMEs that need to be fixed.

Also, needs a good description.  Something like...

Our comparison of files with the same basename is only helpful when we
have exactly one file with a given basename for each of the possible
rename sources and distinations.  When we have multiple files with the
same basename, we do not know which to compare against.

This most frequently comes up when an entire directory is renamed, and
then common filenames (such as 'Makefile' or '.gitignore' or
'build.gradle') that exist within potentially with in every subdirectory
are part of the rename puzzle.  However, when an entire directory is
renamed, (a) we often have several files within that directory and
subdirectories that are renamed without changes, and (b) the original
directory disappeared giving us a hint about when we can apply an extra
heuristic.  So, if there isn't a single basename match, and the
directory in which the file was found has disappeared, use exact renames
of files within the directory to determine where the directory may have
been renamed to.  Take the single most such likely directory, and see
if applying that diretory rename to the file results in a target
filename that is in the potential rename set.  If so, compare it for
similarity, and if it is sufficiently similar, then record the rename.

Signed-off-by: Elijah Newren <newren@gmail.com>