diffcore-rename: attempt to accelerate initialize_rename_guess_info
When there are basename matches but there are more than one, we like to
use directory renames to guide our guessing.  But we only need directory
renames corresponding to relevant_sources and relevant_targets since we
are only looking for renames that match those.

In my testing, this appears to drop overall runtime for basename matches
from a over .6 seconds (across 35 rebases) down to almost exactly .4
seconds; in other words, speeds up overall basename matching by 33%.
(My testing wasn't perfect, and the other memory leak fixes between
could have accounted for part of the difference, but other things
measured about the same as they had before, and a drop of .674 to .404
looked great to me.  Also, this was driven by the fact that some quick
checks showed I had 1629 directories removed, but the number of relevant
sources was usually 1 or 2, and rarely would get up to 218, but any of
those is much less than 1629).

Signed-off-by: Elijah Newren <newren@gmail.com>