diffcore-rename: aborted attempt at faster setup
It turns out that the code has some hard assumptions that the entries in
the queue come in sorted order to diffcore-rename, because nothing after
diffcore-rename does sorting, but there are regression tests that expect
the results in a very specific order.  That means that splitting the queue
into three different ones (not-rename-related, rename_src, and rename_dst)
and then just appending at the end doesn't work.  But adding a QSORT is
more expensive than the lookups that were done at the end before.  So I'm
going to commit my changes as they are for reference, then revert them.

Notes:

  - I got this to pass a bunch of tests, but it fails (at least) on
    break + typechange + copy detection (t4008.10).

There were two basic ideas:
  - writes directly to the output queue the pairs not involved in rename
    detection at the beginning, then at the end only adds to the output
    queue from rename_src and rename_dst.
  - makes rename_dst structure more similar to rename_src -- just needs a
    diff_filepair.  Has to modify the rest of the code for how it finds
    things, though.  This may no longer be necessary given the changes
    to write pairs from rename_dst, because then there is no need to look
    it up without modifying how the data structure.

It also adds a lot of debugging as I'm trying to figure out where I went
wrong, particularly with copies and break detection.