merge-ort: implement trivial tree merges
If a tree's hash on one side of history matches the hash of the
corresponding tree from the merge base, then we can sometimes avoid
traversing into the subdirectory and just take the tree or file on the
other side of history as the resolution.

Doing this "trivial tree merge" is not always safe because the other
side of history may have important rename targets that need to be
included in renaming handling and directory rename handling.  Further,
the source removal might affect directory rename handling for one
directory up.  However, we can cache the matching directory, then
traverse all other paths in the repository.  After we have traversed all
paths other than potential trivial tree merges, if we don't have to
worry about any renames or directory renames (or already have enough
information to handle them in cached_pairs), then we can resolve the
trivial tree merge at the end.  If we need information from the
potential "trivial tree merge" after all, then at the end we simply
recurse into the tree at that point instead of resolving it directly.

Note: We need to clear match_mask when it's set for directories or else
the file resolution logic elsewhere will get confused, particularly with
tests 17, 25, and 30 of t6043 (6b, 7e, and 8e).

Signed-off-by: Elijah Newren <newren@gmail.com>