strmap: attempt to create smarter strmap_clear()
reset_maps() in merge-ort was taking about 12% of overall runtime in my
rebase 35 patches of linux.git across a big rename testcase.  For most
of these, it is freeing all the memory associated with each of the
strmaps, just to immediately allocate a new array again -- and one that
is likely smaller than needed (thus resulting in later need to rehash
things) even though the ending size is likely almost perfectly sized for
the next commit we want to pick.

Add some new API to hashmap to clear a hashmap of entries without
freeing map->table (and instead only zeroing it out like alloc_table()
would do, along with zeroing the count of items in the table).  Make use
of this new API within strmap_clear().  This reduces the cost of
reset_maps() by about 20%, dropping my overall runtime of my rebase
testcase by just under 2%.

Signed-off-by: Elijah Newren <newren@gmail.com>