merge-ort: allow trivial tree merges even when renames are present
renames tend to turn off trivial tree merges because we have to traverse
down into trees to find potential rename targets.  But even if we have
relevant_sources for renames which don't correspond to cached_pairs()
from a previous rebase, the speedup from having fewer paths to deal with
in process_entries() is so great, that we really want to take advantage
of the trivial tree merges.  This suggests a strategy.  Instead of:

    collect_merge_info()
    detect_and_process_renames()
    process_entries()

(each of which are roughly equal in computation cost for the testcase
I'm measuring, by the way), we can instead do the following:

    collect_merge_info()
    detect_and_process_renames()...up to the regular renames, anyway
    if (conditions are met):
        cache the renames
        go back to the collect_merge_info() step
    otherwise:
        finish rename processing
        process_entries()

Due to the caching of renames right after detecting them, the second
time through the detect_regular_renames() becomes a no-op, and the
conditions will not be met the second time which prevents an infinite
loop.  And the benefit kicks in when we call process_entries() with far
fewer paths to handle (and the second collect_merge_info() didn't cost
us nearly as much since it also had far fewer paths to handle).

The "conditions are met" is essentially a comparison of number of paths
in opt->priv->paths to how many that strmap would have had if we had
been able to take advantage of trivial tree merges.  I haven't measured
precisely what that factor should be, but the linux kernel testcases I
was looking at had factors of 50 to 250 and made the optimization very,
very nice; it provided an overall speedup of approximately 20%.

Signed-off-by: Elijah Newren <newren@gmail.com>