t6043: shift test 12b to test_expect_failure
I think this may be okay.  I believe it stems from my optimizations that
I don't care to recurse into a directory if it matches one side, because
"the point of directory rename detection is to take new files within the
directory added on the unrenamed side to the new directory name on the
renamed side".  The directories here match on one side and thus don't
have "any new files within that directory on the unrenamed side".

Thus, I believe this "failure" only occurs on the rename-each-other
situation where no new files are added to either directory.  And when
both sides rename each other's directory, things are already insane so
who cares what the resolution is?

And I certainly don't want to undo those huge performance gains...