merge-ort: fix t6043 tests 6b and 8e
The fix is ugly, and there may be a better way (e.g. checking all pairs
for those that have the same rename source and excluding them from the
directory rename detection -- though then there are questions about
whether that causes a directory rename with just one or two remaining
paths that were renamed elsewhere).  But I'm tired of looking at it, and
this at least brings it in line with merge-recursive.  I'll come back
and look at it later.

Signed-off-by: Elijah Newren <newren@gmail.com>