merge-ort: avoid inappropriate caching of deletes, fix codeflow
cached_irrelevant basically serves similarly to cached_pairs, except
that (a) the former is always for deletes, and (b) we don't actually
know that it was a delete we just stated that we don't care if it was a
rename because we don't need it for _this_ commit.  However, the fact
that we don't need it for this commit does not always guarantee we won't
need it for a subsequent commit; if the next commit we want to pick in a
cherry-pick or rebase modifies the source file on the other side of
history then suddenly we will need that rename.  Fixing this basically
boiled down to adding a couple more lines of code:

    if (content_relevant)
        strset_remove(&renames->cached_irrelevant[side], pathname);

However, these two lines required rewiring to get a bit more information
available so that we could compute content_relevant, and we could also
take advantage of that rewiring to move a bit of rename-specific logic
out of collect_merge_info_callback() and push it into add_pair().

Signed-off-by: Elijah Newren <newren@gmail.com>